refactor(viewer): split CesiumOverlay into the four things it was doing
1,058 lines carrying the Cesium viewer lifecycle, the coordinate bridge, the
model lifecycle and the solar study at once — four subjects with four different
histories, interleaved. CodeRabbit flagged it on #2586; #2593 filed the plan.

Now 419 lines, and it reads as what it is: create the viewer, render the
container, call four hooks in the order their effects used to sit in.

  cesium/cesium-module        the lazy CesiumJS import they share
  cesium/useCesiumBridge      WHERE the model sits — ENU/ECEF, convergence,
                              geoid undulation, terrain clamp, placement drafts
  cesium/useCesiumModel       WHAT is drawn — GLB build, readiness-gated swap,
                              matrix updates
  cesium/useCesiumSolar       lighting, shadows, sun-path dome, sky
  cesium/useCesiumCameraSync  the per-frame camera mirror

The ordering that keeps this behaviour-preserving is the whole risk, so it is
written down rather than assumed: React runs effects in declaration order and
cleanups in reverse, and each hook documents where it must be called and why.
Two teardown paths the viewer effect cannot reach on unmount — the model's and
the solar study's — are now explicit `invalidate()` callbacks instead of
reaching into refs that had moved. The camera loop's rAF handle stays owned by
the caller, because the viewer effect cancels it from its own cleanup, which
runs BEFORE the hook's; without that the loop can tick once against a destroyed
viewer.

Verified end to end on the #2558 model, twice — after the first three
extractions and again after the bridge came out: world view loads, a visibility
change rebuilds the model, the solar study lights the scene and publishes sun
info, a data-source change DESTROYS and recreates the viewer, and the recreated
viewer still rebuilds on the next visibility change. Zero console errors in
every run. Viewer suite 4,102 passing.
